fix: 打通 text 渲染管线（R1 三断点）+ 第二轮基线验收记录
- MBTileDataEmitter: text/poi positions 打包为 Float64Array（原生消费端按
  Float64 解读 buffer，Float32 导致 BufferAttribute.count=0、零 TextElement）
- MBFontCatalogBuilder: registerGlyph 键改用 String(FontStyle.Regular)（查找键
  为数字枚举，"Regular" 字符串永不命中 → replacement glyph 被跳过）
- TileGeometryCreator: symbol-only 瓦片 objects.length===0 时 forceHasGeometry(true)，
  否则 hasGeometry=false 不进 renderedTiles、placeText 永不放置文本元素
- 验证：text-size/default、text-field/literal actual 渲染出文字，管线端到端打通
- docs: §11 第二轮验收记录（204/2646 通过，7.71%，+22）

+++ @flywave/flywave-mbstyle-datasource/src/MBTileDataEmitter.ts
@@ -1119,10 +1119,13 @@ export class MBTileDataEmitter {
         const lineGeoms = this.getLineGeometries();
 
         // Finalize text/POI geometries: convert temp arrays to BufferAttributes.
+        // NOTE: the native consumers (TileGeometryCreator.createTextElements,
+        // PoiManager) reinterpret this buffer as Float64Array, so the positions
+        // MUST be packed as float64 bytes (see VectorTileDataEmitter:1759).
         for (const tg of [...this.m_textGeometries, ...this.m_poiGeometries] as any[]) {
             const positions = (tg as any)._positions as number[] | undefined;
             if (positions && positions.length > 0) {
-                const arr = new Float32Array(positions);
+                const arr = new Float64Array(positions);
                 tg.positions = {
                     name: 'position',
                     buffer: arr.buffer,
